Desktop/プロジェクト/250614_hac_iftool/backend/requirements-cloudrun.txt
# =============================================================================
# Cloud Run用の最適化された依存関係（軽量版）
# =============================================================================

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP関連
httpx==0.25.2
requests==2.31.0

# API ドキュメント
pydantic==2.5.0
pydantic-settings==2.1.0

# 設定・環境変数
python-dotenv==1.0.0

# Google Cloud必須
google-cloud-firestore==2.13.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0

# AI/ML（最小限）
google-generativeai==0.3.2

# CORS対応
python-multipart==0.0.6